web_gateway: carve access_gates.rs (pure move)
Gateway access gates move to web_gateway/access_gates.rs: session
bearer-token minting + verification, HTTP operation classification for
IAM, and per-frame websocket control authorization, with their 19
tests.
